Accept "invalid" field names.
Some of our Avro schemas are generated with "invalid" names and this
library refuses to accept them.  That would be fine if it was only
on writing an Avro schema with such fields but it also refuses to
read any Avro schema, which is highly annoying.

See also https://github.com/linkedin/goavro/issues/286